+++ backend/tp-lib.hpp
@@ -106,11 +106,11 @@ struct Longitude
 // 観光スポットを表す構造体
 struct Place
 {
-    ll id;
+    int id;
     BusinessHours business_hours; // 営業時間（なければ0時00分〜23時59分）
-    ll arrive_before;             // 到着したい時間の上限
-    ll arrive_after;              // 到着したい時間の下限
-    ll stay_time;                 // 滞在時間（分）
+    int arrive_before;             // 到着したい時間の上限
+    int arrive_after;              // 到着したい時間の下限
+    int stay_time;                 // 滞在時間（分）
     int priority;                 // 行きたい度
     double longitude;             // 経度
     double latitude;              // 緯度
@@ -163,7 +163,7 @@ struct StateBase
 {
     Place *place;
     StateBase *prev;
-    double arrive_at;
+    int arrive_at;
     ll visited;
 
     StateBase()
@@ -195,7 +195,7 @@ struct StateBase
     }
 
     // スコア関数
-    virtual double score() const = 0;
+    virtual int score() const = 0;
 };
 
 template <StateBaseType T>
